requirements.txt
discord.py[voice]>=2.7.0
python-dotenv>=1.0.0
yt-dlp>=2026.3.31
PyNaCl>=1.5.0
davey
fastapi>=0.115.0
uvicorn>=0.34.0
jinja2>=3.1.4
aiosqlite>=0.20.0
python-multipart>=0.0.12
